Fix pre-commit for asset compilation missing rich as dependency
After recent refactors, common_pre_commit_utils have rich as
dependency so any pre-commits that cannot have rich set as dependency
should not use it.

The "compile-www-assets" and "compile-www-assets-dev" pre-commits
are "node" pre-commits - which means that they are not installing
python dependencies - they at most might install node dependencies,
and we cannot specify rich as dependency using pre-commit mechanisms.

The "compile-www-assets" and "compile-www-assets-dev" are pretty
special, because they are "manual" stage pre-commits so they are
not executed in CI and we have not detected them failing in
latest refactor.

This PR makes those pre-commit python scripts work only using
stdlib imports and add warning to avoid adding other imports.
